% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_ads_ga_format.R
\name{fb_ads_ga_format}
\alias{fb_ads_ga_format}
\title{fb_ads_ga_format}
\usage{
fb_ads_ga_format(fb_ads_df, utm_medium, utm_source)
}
\arguments{
\item{fb_ads_df}{The data frame of advertisting data from Facebook}

\item{utm_medium}{The medium of the campaign, this should match the values you set in your UTM codes in the URLs in the Facebook ads}

\item{utm_source}{The source of the campaign, this should match the values you set in your UTM codes in the URLs in the Facebook ads}
}
\description{
Format your FB ads data for upload to GA
}
