% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fbads_builder.R
\name{fbads_builder}
\alias{fbads_builder}
\title{fbads_builder}
\usage{
fbads_builder(ad_account_id, token, start_date, end_date)
}
\arguments{
\item{token}{The Facebook app token, generated in the Facebook developer portal}

\item{start_date}{The first date you want data for}

\item{end_date}{The last date you want data for}

\item{ad_account}{The Facebook ad account ID}
}
\description{
Download Facebook ad data
}
