% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/campaign_url.R
\name{campaign_url}
\alias{campaign_url}
\title{campaign_url}
\usage{
campaign_url(url, source, medium = NULL, campaign_name = NULL,
  term = NULL, content = NULL)
}
\arguments{
\item{url}{The URL you want to adorn}

\item{source}{The source of the campaign, typically would be something like 'facebook' or 'adwords'}

\item{medium}{The medium of the campaign, i.e. what type of campaign this is, usually would be something like 'cpc' or 'paid_social' (optional)}

\item{campaign_name}{The name of your campaign (optional)}

\item{term}{A way to identify paid keywords (optional)}

\item{content}{A way to differentiate between different types of marketing content (for example in an A/B test) (optional)}
}
\description{
Format your URL with UTM tracking codes
}
