% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fb_ads_ga_upload.R
\name{fb_ads_ga_upload}
\alias{fb_ads_ga_upload}
\title{fb_ads_ga_upload}
\usage{
fb_ads_ga_upload(fb_df, ga_property_id, dataset_name)
}
\arguments{
\item{fb_df}{The formatted data frame of advertisting data from Facebook, ready to upload}

\item{ga_property_id}{Your web property ID from Facebook; it should be formatted something like 'UA-'}

\item{dataset_name}{The name of your dataset}
}
\description{
Upload your data to GA
}
